Lab06: Lexer and parser class created
Add basic data types, token to contain the tokenized value from the
lexer. The parser class is empty and contain no implementation. enum
class statement and expression might be changed depending on if there's
a better way to structure it.

Disabled /fsanitizer=address flag on Windows, this flag only applies
when working in Visual Studio.