Lab05: All test case passes
The memory growth needs to be explored and have another implementations
later.

The plus operator interface was trying to return a reference which is
not correct, the test case now checks if the left operand is modified
and then check the assigned value if it got the correct new size.

The string size now returns the size of string without null termination
character. Before it was included.