Lab06: string literal and unary expression
Fixes token::is_identifier regex, this was not a valid range.

Move sanitize_str lambda to static member function in token. This
function is later used in recursive print to sanitize string output. Not
all output is sanitized yet in str member function for ast_nodes.

Check for eof in lexer with string literal.

Default file name to unknown.